+++ RedPandaIDE/src/widgets/searchresultview.h
@@ -45,62 +45,61 @@ struct SearchResultTreeItem {
     bool selected;
 };
 
-struct SearchResults{
-    QSynedit::SearchOptions options;
-    bool useRegex;
-    QString keyword;
-    QString statementFullname;
-    SearchFileScope scope;
-    SearchType searchType;
-    QString filename;
-    QString folder;
-    QString filters;
-    bool searchSubfolders;
-    QList<PSearchResultTreeItem> results;
+
+class SearchResultModel;
+
+class SearchResults{
+public:
+    explicit SearchResults(const QString &keyword, QSynedit::SearchOptions options,
+                           bool useRegex, SearchFileScope scope, SearchType type,
+                           const QString& folder, const QString& filters,
+                           bool searchSubFolders);
+    explicit SearchResults(const QString& keyword,
+                           const QString& symbolFullname,
+                           const QString& filename,
+                           SearchType searchType,
+                           SearchFileScope scope);
+    const QSynedit::SearchOptions &options() const;
+    void setOptions(const QSynedit::SearchOptions &newOptions);
+
+    bool useRegex() const;
+    void setUseRegex(bool newUseRegex);
+    const QString &keyword() const;
+    void setKeyword(const QString &newKeyword);
+    SearchFileScope scope() const;
+    void setScope(SearchFileScope newScope);
+    SearchType searchType() const;
+    void setSearchType(SearchType newSearchType);
+    const QString &filename() const;
+    void setFilename(const QString &newFilename);
+    const QString &folder() const;
+    void setFolder(const QString &newFolder);
+    const QString &filters() const;
+    void setFilters(const QString &newFilters);
+    bool searchSubfolders() const;
+    void setSearchSubfolders(bool newSearchSubfolders);
+    const QList<PSearchResultTreeItem> &results() const;
+    PSearchResultTreeItem result(int idx) const;
+    const QString &symbolFullname() const;
+    void setSymbolFullname(const QString &newSymbolFullname);
+
+private:
+    QSynedit::SearchOptions mOptions;
+    bool mUseRegex;
+    QString mKeyword;
+    QString mSymbolFullname;
+    SearchFileScope mScope;
+    SearchType mSearchType;
+    QString mFilename;
+    QString mFolder;
+    QString mFilters;
+    bool mSearchSubfolders;
+    QList<PSearchResultTreeItem> mResults;
+
+    friend SearchResultModel;
 };
 
 using PSearchResults = std::shared_ptr<SearchResults>;
-
-class SearchResultModel : public QObject {
-    Q_OBJECT
-public:
-    explicit SearchResultModel(QObject* parent=nullptr);
-    PSearchResults addSearchResults(const QString& keyword,QSynedit::SearchOptions options,
-                                    bool useRegex,
-                                    SearchFileScope scope, const QString& folder=QString(), const QString& filters=QString(), bool searchSubfolders=true);
-    PSearchResults addSearchResults(
-            const QString& keyword,
-            const QString& symbolFullname,
-            SearchFileScope scope);
-    PSearchResults results(int index);
-    void notifySearchResultsUpdated();
-    int currentIndex() const;
-    int resultsCount() const;
-    PSearchResults currentResults();
-    void setCurrentIndex(int index);
-    void clear();
-    void removeSearchResults(int index);
-signals:
-    void modelChanged();
-    void currentChanged(int index);
-private:
-    QList<PSearchResults> mSearchResults;
-    int mCurrentIndex;
-
-};
-
-class SearchResultListModel: public QAbstractListModel {
-Q_OBJECT
-    // QAbstractItemModel interface
-public:
-    explicit SearchResultListModel(SearchResultModel* model,QObject* parent=nullptr);
-    int rowCount(const QModelIndex &parent) const override;
-    QVariant data(const QModelIndex &index, int role) const override;
-public slots:
-    void onResultModelChanged();
-private:
-    SearchResultModel *mSearchResultModel;
-};
 
 class SearchResultTreeModel : public QAbstractItemModel {
 Q_OBJECT
@@ -121,12 +120,13 @@ public:
     bool selectable() const;
     void setSelectable(bool newSelectable);
 
-public slots:
-    void onResultModelChanged();
 private:
     SearchResultModel *mSearchResultModel;
     bool mSelectable;
-
+private:
+    void resetModel();
+    void beginAddItem();
+    void endAddItem();
     // QAbstractItemModel interface
 public:
     Qt::ItemFlags flags(const QModelIndex &index) const override;
@@ -134,9 +134,39 @@ public:
     // QAbstractItemModel interface
 public:
     bool setData(const QModelIndex &index, const QVariant &value, int role) override;
+friend class SearchResultModel;
 };
 
-using PSearchResultTreeModel = std::shared_ptr<SearchResultTreeModel>;
+class SearchResultModel : public QAbstractListModel {
+    Q_OBJECT
+public:
+    explicit SearchResultModel(QObject* parent=nullptr);
+    PSearchResults addSearchResults(const QString& keyword,QSynedit::SearchOptions options,
+                                    bool useRegex,
+                                    SearchFileScope scope, const QString& folder=QString(), const QString& filters=QString(), bool searchSubfolders=true);
+    PSearchResults addSearchResults(
+            const QString& keyword,
+            const QString& symbolFullname,
+            SearchFileScope scope);
+    PSearchResults results(int index);
+    int currentIndex() const;
+    int resultsCount() const;
+    PSearchResults currentResults();
+    void addResultToSearchResults(PSearchResults results, PSearchResultTreeItem item);
+    void setCurrentIndex(int index);
+    void clear();
+    void removeSearchResults(int index);
+    int rowCount(const QModelIndex &parent) const override;
+    QVariant data(const QModelIndex &index, int role) const override;
+    SearchResultTreeModel *treeModel() const;
+signals:
+    void currentIndexChanged();
+private:
+    QList<PSearchResults> mSearchResults;
+    int mCurrentIndex;
+    SearchResultTreeModel *mTreeModel;
+};
+
 
 class SearchResultTreeViewDelegate: public QStyledItemDelegate{
 Q_OBJECT